Fixed a bug with Schema interactors later in the chain
What happened:
When an interactor chain starts without a schema, but is assigned one
later, it will call Schema::Context.build instead of Context.build

`OpenStruct`s respond to all missing methods with `nil` by default. In
this case, `select` is not a method on an `OpenStruct`.

By calling `to_h` on an OpenStruct we are able to select the attributes
that are included in the schema